Time: 1 ms (100%), Space: 62.3 MB (51.08%) - LeetHub

+++ LeetCode/Medium/3689-maximum-total-subarray-value-i/3689-maximum-total-subarray-value-i.java
@@ -1,8 +1,8 @@
 class Solution {
     public long maxTotalValue(int[] nums, int k) {
         int n = nums.length;
-        long max = Integer.MIN_VALUE;
-        long min = Integer.MAX_VALUE;
+        int max = Integer.MIN_VALUE;
+        int min = Integer.MAX_VALUE;
 
         for(int i = 0 ; i< n ;i++){
             max = Math.max(nums[i] , max);
